# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SRM API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/{TenantId}/roles/{RoleId}/permissions": {
            "get": {
                "tags": [
                    "RolesPermissions"
                ],
                "summary": "Get list of  permissions by role ID",
                "description": "Returns list of permissions by role id ",
                "operationId": "getRolesPermissionsList",
                "parameters": [
                    {
                        "name": "TenantId",
                        "in": "path",
                        "description": "Tenant id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "RoleId",
                        "in": "path",
                        "description": "Role's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched role's permissions"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "RolesPermissions"
                ],
                "summary": "Create permissions by role ID",
                "description": "Create permissions by role ID",
                "operationId": "createRolesPermissions",
                "parameters": [
                    {
                        "name": "TenantId",
                        "in": "path",
                        "description": "Tenant id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "RoleId",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "permissions"
                                ],
                                "properties": {
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully created role's permissions"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "basicAuth": {
                "in": "header",
                "name": "bearerAuth",
                "type": "HTTP",
                "scheme": "basic"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}